requirements.txt
# Core Dependencies
python-dotenv==1.0.0

# Google APIs
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Jira
jira==3.6.0

# AI/LLM
anthropic==0.18.1
openai==1.12.0

# Scheduling
schedule==1.2.1
APScheduler==3.10.4

# Data Processing
PyYAML==6.0.1
requests==2.31.0

# Database
sqlalchemy==2.0.25

# Utilities
python-dateutil==2.8.2
pytz==2024.1
